+++ src/k8s_sandbox/_helm.py
@@ -172,12 +172,28 @@ async def uninstall(release_name: str, namespace: str, quiet: bool) -> None:
 
     The number of concurrent uninstall operations is limited by a semaphore.
 
+    "Release not found" errors are ignored.
+
     Args:
         release_name: The name of the Helm release to uninstall (e.g. abcdefgh).
         namespace: The Kubernetes namespace in which the release is installed.
         quiet: If False, allow the output of the `helm uninstall` command to be written
           to this process's stdout/stderr. If True, suppress the output.
     """
+
+    def _is_release_not_found_error(stderr: str) -> bool:
+        # The consequence of a false positive is to discard a useful error message, so
+        # err on the side of strictness.
+        return (
+            re.match(
+                rf"^Error: uninstall: Release not loaded: {release_name}: release: not "
+                "found$",
+                stderr,
+                re.IGNORECASE,
+            )
+            is not None
+        )
+
     async with _uninstall_semaphore():
         with inspect_trace_action(
             "K8s uninstall Helm chart", release=release_name, namespace=namespace
@@ -195,14 +211,21 @@ async def uninstall(release_name: str, namespace: str, quiet: bool) -> None:
                 ],
                 capture_output=True,
             )
-            if not quiet:
+            # A helm uninstall failure with "release not found" implies that the release
+            # was never successfully installed or has already been uninstalled.
+            # When a helm release fails to install (or the user cancels the eval), this
+            # uninstall function will still be called, so these errors are common and
+            # result in error desensitisation.
+            is_release_not_found_error = _is_release_not_found_error(result.stderr)
+            if not quiet and not is_release_not_found_error:
                 sys.stdout.write(result.stdout)
                 sys.stderr.write(result.stderr)
-            if not result.success:
+            if not result.success and not is_release_not_found_error:
                 _raise_runtime_error(
                     "Helm uninstall failed.",
                     release=release_name,
                     namespace=namespace,
+                    returncode=result.returncode,
                     stdout=result.stdout,
                     stderr=result.stderr,
                 )